Common checks refactored
