+++ github/ci/services/prometheus-stack/e2e/BUILD.bazel
@@ -4,13 +4,9 @@ go_test(
     name = "go_default_test",
     srcs = ["e2e_test.go"],
     deps = [
-        "//github/ci/services/common/k8s/pkg/client:go_default_library",
-        "//github/ci/services/common/k8s/pkg/portforwarder:go_default_library",
-        "//github/ci/services/common/k8s/pkg/wait:go_default_library",
+        "//github/ci/services/common/k8s/pkg/check:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/ginkgo/extensions/table:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes:go_default_library",
     ],
 )
