+++ github/ci/services/sippy/e2e/BUILD.bazel
@@ -4,8 +4,7 @@ go_test(
     name = "go_default_test",
     srcs = ["e2e_test.go"],
     deps = [
-        "//github/ci/services/common/k8s/pkg/portforwarder:go_default_library",
-        "//github/ci/services/common/k8s/pkg/wait:go_default_library",
+        "//github/ci/services/common/k8s/pkg/check:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
     ],
